docker/requirements.txt
boto3
pypdf
pdf2image
Pillow
python-multipart
python-magic